More changes
Character on screen and moves - all computer characters/monsters drawn;
Bugs: when character moves, items on map move too? MAP_WIDTH set to 20
but really is 40  because of the way that characters are displayed ->
this messes up the IsPassable function because characters can only walk
halfway across map. But, when I made it so that they could walk all the
way, I started getting a segfault